Remove Rc from pipeline commands
I used to think commands have to be shared by using Rc like the original
yash implementation. The purpose of the reference counting is to protect
the function body from being dropped while it is being executed. For
that purpose, it is more correct to share function bodies rather than
sharing commands contained in function bodies.