Don't clone System when cloning Env
Cloning an Env instance should implement the semantics of creating a
subshell from an existing shell execution environment. However, the
previous implementation did not match that in that the System instance
was being cloned as if the entire operating system environment is
cloned. In the correct implementation, the subshell should share the
System with the parent.

This commit removes System::clone_box and changes Env::system from
Box<dyn System> to Rc<RefCell<dyn System>>.