Remote ExpandToField::expand_to_fields
An async_trait method returns the result in a boxed future. We should
avoid the boxing overhead by not implementing such a simple wrapper
of another async function.

The function could be moved to be a module-level function, but this
commit removes it as it is not used from anywhere.