Pass job control status to subshell task
The subshell task should be able to tell the actual job control status
of itself. This commit adds a second parameter to the task to make it
possible. `yash_semantics::command::item::execute_async` now uses the
new parameter, which removes a redundant call to `Env::controls_jobs`
before starting the subshell.