Ignore SIGINT & SIGQUIT in async item
POSIX requires that asynchronous items that are not job-controlled
ignore SIGINT and SIGQUIT. This commit implements that behavior.

In this commit, I add the MaskGuard structure, which I use in the
Subshell::start function to remember the original signal-blocking mask
and temporarily block SIGINT and SIGQUIT while forking a child process.
I call System::sigmask directly from MaskGuard, despite the signal
handler controlling mechanisms in TrapSet and SelectSystem, to keep the
implementation simple. If we let TrapSet control the mask, we would need
to extend SignalSystem with a method to block signals, making
SelectSystem more complicated. Such a new public API function would have
to be generic enough to support any combination of API calls, including
those that were unnecessary to achieve our original goal of handling
SIGINT and SIGQUIT.