Amend yash-quote documentation
The previous doc test was not performing the expected comparison because
the PartialEq implementation for Cow only examines the contained values
regardless of whether they are borrowed or owned. Ideally, we should
demonstrate that the quote function returns a predictable variant of Cow
for each example, but we need to wait for assert_matches to be
stabilized to do so. In the meanwhile, we just compare contained values.