Add yash-semantics to yash-builtin as optional dependency
The exec built-in will need to depend on the command search implemented
in the yash-semantics crate, so the yash-builtin crate needs it as a
dependency. We make it an optional dependency so that users can avoid
dragging it in when they only want built-in implementations that do not
depend on it.

Fixes https://github.com/magicant/yash-rs/issues/147