exec: Invoke external utility
This commit implements the most basic functionality of the exec built-in.
It invokes an external utility when given an operand.

To expose some utility functions to the exec built-in, this commit also
reorganizes public items in the yash-semantics crate. The Command trait
is now at yash_semantics::command::Command, and
yash_semantics::command::simple_command is now public.