Return specified exit status in Divert
After this commit, the exit built-in always returns the specified exit
status in Divert::Exit. The built-in's exit status is always taken from
the previous command's exit status.

This change should make it easier to implement the correct exit status
for the case where the exit built-in is invoked without an operand in a
trap handler.