Always restore exit status after running traps
The run_traps_for_caught_signals function now restores the exit status
even if the trap action results in a divert.

If the trap action contains an exit command without an exit status
(which yields `Divert::Exit(None)`), the shell should exit with the exit
status of the last command executed before entering the trap. To achieve
this behavior, the function has to restore the exit status after the
trap action.