Reorganize break/continue built-in implementation
This commit splits the break/continue built-in implementation into the
syntax part, semantics part, and the main entry point (controller). This
refactoring should make the built-in implementations minimal and more
reusable.

The integration tests for the break built-in is reimplemented in
yash/tests/builtin/break.rs, but it will eventually be replaced with
scripted tests imported from yash.