Replace Variable with PositionalParams in Context
Using Variable for positional parameters was a hack. It is not a
variable, but we used it as a variable because it was convenient. That
made the code unintuitive and incurred unnecessary overhead.

This commit replaces Variable with PositionalParams in Context so that
VariableSet::positional_params(_mut) returns a reference to the distinct
type that provides the correct view of positional parameters.