Report read-only variable name in error message
The refactoring in commit 8e2bcacbe3 (#314) made the error message
for assigning to a read-only variable less informative by removing the
variable name. This commit restores the variable name in the error
message.

To this end, this commit introduces a new error type
`AssignReadOnlyError` that is used to report the error in the
expansion phase. The error type is converted from `AssignError` to
convey the variable name, which is not included in `AssignError`.