Rename builtin doc sections
This commit renames the Syntax sections of builtin doc comments to
Synopsis. This matches the terminology used in POSIX and the man pages
of common utilities.

This commit also renames the Semantics sections to Description. This
also matches the POSIX terminology and more accurately describes the
contents of the sections. The Options and Operands sections describe the
syntax and semantics of the options and operands of the builtins, so it
felt a bit awkward to have separate sections named Syntax and Semantics.