Don't omit export or readonly commands when printing arrays or functions
This commit introduces a new field `builtin_is_significant` to the
`PrintContext` struct. When this field is true, the command that invokes
the built-in is always printed regardless of the attributes of the
variables or functions. This field is true for the export and readonly
built-ins, but false for the typeset built-in.