Faster test for long here-document
The previous test for long here-document was very slow because it was
using the read built-in to read the output of cat in a loop, where the
built-in reads a line byte by byte. This commit replaces that loop with
a single invocation of diff with a prepared expected output.

This commit also adds a test for a parameter expansion in the long
here-document.