Wake ProcessRunner when process is resumed
The previous implementation of ProcessRunner was not woken up when the
process was resumed since ProcessRunner was waiting on SignalStatus and
SIGCONT is not necessarily caught by the process. This commit makes
sure that ProcessRunner is woken up whenever the process is resumed by
registering a waker to the process.