Redefine Pid independently of nix
Previously, the Pid type was re-exported from the nix crate, which may
possibly introduce backward incompatible changes in the future. For the
maximum semantic versioning compliance, we should define the type in our
own crate.

This commit replaces yash_env::job::Pid with a newly defined structure,
which provides similar API (if not the same) compared with the nix
version.