fg: Remove job if job has already finished
If a job is finished, the `fg` built-in should remove the job from the
job set. This commit fixes the bug where the built-in does not remove
the job if the job has already finished. Previously, the built-in was
removing the job only if the job finished after the built-in resumed
the job.