Test unused_crate_dependencies in CI
To make sure we don't have any unnecessary dependencies in Cargo.toml,
this commit adds a command line that checks with the
unused_crate_dependencies lint.

This commit also adds command lines that build with all combinations of
features to catch possible build errors that do not happen with the
check command.

Some test commands are removed from check-extra.sh because they are
already run in check.sh.