Return Expansion from PathEnv::path
When the command built-in is invoked with the -p option, it uses the
standard path to search for the command. To accommodate the dynamic
retrieval of the path from the underlying system, the PathEnv::path
function should return an Expansion instead of a reference to a variable.
The Expansion can contain an arbitrary owned string, allowing the
function to return a result that doesn't reference a value retained in
the environment.