Retain Times values in f64
This change makes the `Times` struct contain seconds in f64 instead of
clock ticks in u64.

The `tms` struct returned by the underlying `times` function contains
values in `clock_t`. The previous implementation assumed that it is an
integral type, but POSIX allows it to be a floating-point type. This
change makes the `Times` struct use f64 for all time values, which
can be cast from any types `clock_t` may be.