umask: InvalidNumericMode and InvalidSymbolicMode
The error InvalidMode is split into InvalidNumericMode and
InvalidSymbolicMode. The former is used when an operand starts with a
digit and is not a valid mode. The latter is used when an operand does
not start with a digit and is not a valid mode.

With the two error types, the error message can be more specific.