Make $TEST_NO available to test scripts
In the test_symbolic_operand function, we need a unique number to create
a directory name. The caller passes $LINENO as the first argument, but
it is not supported by older shells. To pass the line number to the test
script, we now export the $TEST_NO environment variable in the test
harness.