ulimit: Support hard and soft operands
This commit adds support for the `hard` and `soft` operands to the
`ulimit` built-in. The new enum `SetLimitValue` represents the operand
value, which can be a numeric value, `unlimited`, `soft`, or `hard`.

The check for overflow is now done in the `set` module instead of the
`syntax` module, so that the `syntax` module can be focused on parsing
command-line arguments in a purely syntactical manner.