Substitute alias after line continuation after blank-ending alias
The next token after a blank-ending alias should be subject to alias
substitution. However, the shell was not substituting the alias if there
was a line continuation before the next token. This was because the
LexerCore::is_after_blank_ending_alias function was only looking for the
last non-blank character before the current position. This commit
changes the function to also check if the character is a line
continuation.

To enable this, a new LexerCore::mark_line_continuation function is
added to mark the characters in a given range as line continuation. The
Lexer::line_continuation function now calls this function to mark the
characters it has consumed as line continuation.