Correct source character location indexes after alias substitution
The previous implementation of the LexerCore::peek_char method naively
used the current index to calculate the start index of the location at
the current position. This approach was incorrect because the index
could be changed by an alias substitution that changed the length of the
source code. This commit fixes the issue by introducing a new method,
LexerCore::next_index, that computes the index from the last character
of the source code.

Fixes https://github.com/magicant/yash-rs/issues/280