Clear CLOEXEC flag on FDs for temporary files
All `System::open_tmpfile` implementations should consistently return a
file descriptor without the `O_CLOEXEC` flag set. Otherwise, it would
break callers that expect the file descriptor to be visible to the user.

Previously, the 'command exec retains redirection' test was failing
because the file descriptor to the temporary file containing the here-
document content was not visible to the command.