gropagate exit status of shell error in EXIT trap
This commit applies the same fix as in commit 5e82a55b to the EXIT trap.
If the EXIT trap is interrupted by a shell error, the exit status of the
shell error is propagated to the calling context, which means that the
exit status of the shell error overrides the exit status of the shell.

The `run_trap` function is refactored to be used by both the EXIT and
signal traps. The `run_trap` function now takes a `Condition` argument
to specify the type of the trap, which is no longer limited to signals.