Remove Result from closure parameter of Subshell::new
This commit refactors the closure parameter of Subshell::new to return
a Future that resolves to () instead of Result. The Subshell
implementation no longer calls Env::apply_result on the result of the
closure. Instead, Env::apply_result must be called explicitly in the
closure, if necessary.

This commit is a preparation for the upcoming fix for the issue that
the EXIT trap is not run in subshells. The EXIT trap should be run after
Env::apply_result is called in the closure, which cannot be implemented
in the yash-env crate as trap execution depends on the yash-semantics
crate.