Compute shell path in System
This commit adds a new method `System::shell_path` that returns the path
to the shell executable. The method encapsulates the logic to find the
shell executable, which is currently implemented as follows:

- On Linux and Android, it returns `/proc/self/exe`.
- On other platforms, it searches for an executable `sh` from the
  default PATH returned by `System::confstr_path`.

The method is used in `command::simple_command::replace_current_process`
to find the shell executable when it needs to fall back to the shell.
The previous logic to find the shell executable is removed from the
function.