Replace SigSet with Signal array in System
This commit replaces the SigSet type with an array of Signal in the
System trait. This change is necessary because the SigSet type is
specific to the nix crate, and the System trait should not depend on
specific types from external crates.

This commit also updates the implementations of the System trait to
reflect this change. SelectSystem::wait_mask is now an optional Vec of
Signal, and virtual Process now uses a BTreeSet of Signal instead of
SigSet.