Use signal::Number in Process::signal_handlings
This commit modifies the type of the `signal_handlings` field of
`Process` from `HashMap<Signal, SignalHandling>` to
`HashMap<signal::Number, SignalHandling>`. This change is necessary
because the `Signal` type is being deprecated in favor of
`signal::Number`.

This commit also modifies the `signal_handling` and
`set_signal_handling` methods of `Process` to operate on
`signal::Number` instead of `Signal`.