Use new signal types in kill built-in
This is part of the ongoing reconstruction to eliminate the use of nix
types in the built-ins. This commit introduces the new signal types that
are now used in the kill built-in. The `Signal` type represents a signal
to be sent or printed by the kill built-in. The validity of the signal
name or number is now checked at the time of execution by the system.