Remove yash_env::trap::Signal
This type was used to represent signals in the trap module, but it has
been replaced by yash_env::signal::Signal. This commit removes the old
type.

This commit is the final step in the extensive restructuring to remove
the nix-based Signal type from the public API of the yash_env crate.
Note that internal uses of the type still remain in the implementation,
which we plan to address later.