Don't break/continue out of a trap
An example in the Rationale section of the POSIX standard for the break
built-in states that a loop does not lexically enclose the built-in if
the built-in is inside a trap that is executed during the loop. This
commit modifies the `Stack::loop_count` method so that the break and
continue built-ins refuse to operate on such a loop.