Update annotate-snippets to 0.11.4
This update of annotate-snippets includes breaking changes in a wide
range of APIs. The most notable is that we can no longer directly access
the fields of `Message` (formerly `Snippet`), which means that we cannot
examine the state of `Message` without rendering it to the final text in
the tests. We decided to remove the corresponding tests, since comparing
the texts would depend on the formatting details of annotate-snippets
and would be much more unstable for us.