Migrate from ReadEvalLoop to read_eval_loop
This commit migrates the codebase from the ReadEvalLoop struct to the
read_eval_loop function. The read_eval_loop function is a top-level
function that allows sharing the environment between the input object
used in the lexer and the command execution functions.

This change finishes the preparation for the upcoming command prompt
support, which will be implemented as another input decorator that
shares the environment with the read-eval loop.