Extract yash-env-test-helper crate
There used to be the same helper functions in some of the crates in the
workspace, but they were duplicated and not shared. This commit
extracts the helper functions into a new crate, yash-env-test-helper, so
that they can be shared among the crates.

The yash-env crate still has duplicated functions because the crate
cannot depend on the helper crate which in turn depends on the yash-env
crate.