Guarantee line number counting with large character index
When you want to get the line number for the last character of the code,
counting the number of characters in the code just to create a correct
Location object is a waste of time. The existing implementation of the
`Variable::expand` function allows the `location.range.start` to be
greater than the character count of the code, in which case the line
number is calculated as if the range were the end of the code. This
commit guarantees this behavior by adding a comment to the function
documentation.