Wrap Source in Rc in yash_syntax::source::Code::source
A new Code instance is created every time the lexer flushes its buffer,
but the Source object is the same for all the Code instances created
from the same input. This change makes the Source object an Rc<Source>
instead of Source to avoid cloning the Source object when creating a new
Code instance.