Attach Vacancy to NonassignableError
The `NonassignableError` enum has been replaced with a struct of the
same name so that it can have a `Vacancy` field. This change is made to
provide an additional annotation for the error message that describes
the state of the parameter value that caused an attempt to the
assignment.

The ErrorCause::related_location method has been deprecated in favor of
the new additional_message method. The new method can return a message
without a location, indicating that the message should be annotated at
the same location as the main message.