Split yash-cli CHANGELOG
The CHANGELOG.md file in the yash-cli crate is split into two files:
CHANGELOG-lib.md for changes to the library crate and CHANGELOG-bin.md
for changes to the shell binary as a whole. Since most shell users are
unlikely to be interested in the library crate changes, we provide
separate files to make it easier to find the relevant information.