Report vacancy on assignment error in parameter expansion
The vacancy is now reported not only for the `NonassignableError` but
also for `AssignReadOnlyError`. This change is made to provide more
information about the error.

I chose to revert the addition of the `additional_message` method to the
`ErrorCause` enum because it could not return more than one message. The
annotations are now directly constructed in `impl MessageBase for Error`
based on the error cause.