Rename Param to BracedParam
In POSIX terminology, a parameter refers to an abstract entity that
can be expanded to a value. The syntactic construct `${...}` is called
a parameter expansion, and should not simply be called a parameter.
Since there are two forms of parameter expansions, the one enclosed
in braces is now called `BracedParam` to avoid confusion.