Use Param in BracedParam
This commit replaces the `name: String` field of the BracedParam struct
with a `param: Param` field. The Param struct contains the precomputed
parameter type in addition to the identifier.

To compute the parameter type correctly, the lexer is now more strict
about acceptable parameters. Specifically, a parameter that starts with
a digit but contains non-digit characters is now considered invalid. The
new variant `SyntaxError::InvalidParam` is returned in this case.