Run initialization file ($ENV) on startup
This change adds support for running the initialization file specified
by the ENV environment variable on shell startup. The initialization
file is executed if the shell is interactive. The initialization file is
executed before the main read-eval loop starts.

The init_file module in the yash-cli crate contains items for running
the initialization files.

The scripted test startup-y.sh is imported from the yash shell:
https://github.com/magicant/yash/blob/3f6f03967a34ccfc53a04b87862f9664b3e98851/tests/startup-y.tst