Don't export IFS outside testcase function
This commit back-ports the fix from yash:
https://github.com/magicant/yash/commit/89f5387622a9d9504641ee81b19fcf78edddaed3

The run-test.sh script depends on field splitting to pass the correct
option to the diff utility in the testcase function. This means the
function should be called with the default IFS. This commit modifies the
test case that needs a non-default IFS to export it only inside the test
case script.