Rename bin_main to main and let it return !
As Rust 1.79.0 allows the main function to be imported from a library
crate, we rename the bin_main function to main and import it in the
binary crate.

The function now calls std::process::exit instead of returning the exit
status. This change makes the function compatible for the binary crate's
main function. It also is a bit more efficient as it does not need to
drop the Env struct before exiting.