Remove the `yash_env::system::fd_set` module
The implementation of the `FdSet` type involved unreliable use of
`MaybeUninit::assume_init` because the actual definition of `fd_set` may
contain padding bits that are not initialized by `FD_ZERO`. This is
unsafe and may lead to undefined behavior, so the `FdSet` type is
removed.

The conversion from `FdSet` to `libc::fd_set` is now done directly in
the `RealSystem` implementation of `System::select`.