Use libc functions for file descriptor operations
This change replaces the nix crate with libc for file descriptor
operations. This change is made to reduce the dependency on the nix
crate.

Since version 0.28.0, the nix crate has started to use the
std::os::fd::AsFd trait, which cannot be used with the file descriptor
wrapper type Fd. This change is necessary to avoid the use of the
AsFd trait and to upgrade the nix crate to version 0.29.0.

This change also modifies the System::isatty method to return a bool
instead of a system::Result<bool>. This change is made to avoid
returning false errors obtained from errno since POSIX does not
require the isatty function to set errno.