Ignore the `noexec` option in interactive shells
The pipeline execution now ignores the `noexec` option in interactive
shells. Previously, the `<yash_syntax::syntax::Pipeline as
command::Command>::execute` method skipped the execution of the pipeline
if the `Exec` shell option was off. Now, it skips the execution only if
the `Exec` and `Interactive` shell options are both off. This change
allows you to exit the shell in interactive mode even if the `noexec`
and `ignoreeof` options are set. (Though the `ignoreeof` option is not
implemented yet.)

The new scripted test `option-y.sh` is adapted from yash's test file:
https://github.com/magicant/yash/blob/703f0b9ec0e959e04289ae75783edf589d3dcc0d/tests/option-y.tst