Remove SendError
This commit removes the `SendError` type from the `forwarder` module.
The `SendError` type was used to indicate errors that occur when sending
a value fails. However, the `SendError` type was not very meaningful
because it was difficult to distinguish between the two cases where the
receiver has been dropped and the value has already been sent. If the
receiver is dropped after it receives the value, the sender will return
`ReceiverDropped`, where `AlreadySent` would be more appropriate.
Therefore, this commit removes the `SendError` type and changes the
`Sender::send` method to return the value back to the caller without any
further information if the receiver has been dropped.