check-msrv: Autodetect expected MSRV
This commit updates the `check-msrv` script to autodetect the expected
minimum supported Rust version (MSRV) for each crate. This is done by
querying the `rust_version` field of the package metadata using `jq`.
This ensures that the script always tests the correct MSRV for each
crate, even if the MSRV is updated in the future.