Parse a terminator token in Parser::case_item
Previously, we parsed the trailing double semicolon in a case item
outside of the `case_item` method. This commit changes the method to
consume the token itself, and returns whether it consumed a terminator
token. This will allow us to parse the case items more naturally and
also to parse different terminators in the `case_item` method in the
future.