Redefine MaybeLiteral::extend_if_literal as extend_literal
The existing implementation of `MaybeLiteral::extend_if_literal` for
`[T]` was wrong in that it returned `Err(result)` where `result` was
modified. This is a violation of the contract of the method, which
should return `result` unmodified in case of failure. To fix this, the
method is redefined as `extend_literal`, which is declared to modify
`result` regardless of success or failure.

The new method is defined to take a mutable reference to an
`Extend<char>` object, instead of an ownership of it, because the
ownership is not necessary for the method to work.