Implement FromStr for EscapedString
To implement FromStr for EscapedString, we need to parse a whole string
as an escaped string, ignoring the surrounding quotes. To this end, this
commit extracts the core logic of parsing an escaped string into a
separate method, Lexer::escaped_string, which takes a closure to
determine the delimiter. The existing method
Lexer::single_quoted_escaped_string is updated to use this new method.