Parser Config
This commit adds the Config struct to the parser module. Config is a
builder for constructing a parser. It allows the user to optionally
customize the parser settings before creating a parser while providing
default settings in case the user does not want to customize them.

This change should improve forward compatibility with settings that may
be added in the future.