Make typeset a declaration utility
This commit makes the typeset built-in a declaration utility as an
extension to the POSIX standard. To make it possible, the read-eval-loop
now configures the parser with the environment as a glossary for
declaration utilities. The glossary implementation looks up the command
name in the built-in utilities and returns the value of the
`is_declaration_utility` field, which is also added in this commit.