Remove yash-cli library from semantic versioning policy
The `yash-cli` package contains both the `yash3` binary and the
`yash-cli` library crate. Following semantic versioning for the both
crates is not practical because we sometimes need to make incompatible
changes to the library crate without breaking the shell's features.
Since the library crate is not intended to be used by other programs,
we will no longer follow semantic versioning for it.