Use pointer::write to initialize raw struct
This commit partially reverts dc878f2d.

The previous commit used the assignment operator to initialize the
`sa_flags` field of the `sigaction` struct, but the operator drops the
previous value of the field, which is not yet actually initialized. That
won't cause any problem in this case since the field is of raw integer
type that does not implement `Drop`, but it is better to use the
`pointer::write` method to initialize any field.