Retain TrapState in GrandState
This commit changes the `GrandState` struct to retain the `TrapState`
struct instead of the `Setting` enum as the current and parent trap
state.

The current definition of `Setting` does not use `Action::Default` or
`Action::Ignore` variants for the initial disposition, which complicates
the computation of the current signal disposition. This commit changes
the `GrandState` struct to use the `TrapState` struct directly, so that
the current signal disposition can always be computed from the `action`
field of the `TrapState` struct.

The `Setting` enum will be removed in the next commit.