Set TrapState::origin to Subshell when entering a subshell
This commit changes the origin of the current trap state to Subshell
when entering a subshell. Previously, the origin was set to Inherited
for backward compatibility but was incorrect. This commit also updates
the tests to reflect the new behavior. Especially, TrapSet::get_state
now returns a reference to the current state instead of None after
entering a subshell.