Reflect EnterSubshellOption::Ignore to current_state.action
Previously, the current state's action was set to Action::Default when
EnterSubshellOption::Ignore was passed to GrandState::enter_subshell.
This commit changes it to Action::Ignore, so that the action indicates
the actual configuration of the trap.

This change is necessary for the trap built-in to print all traps
correctly, including ones that are not configured by the user.