Replace GrandState::get_state with current_state and parent_state
This commit replaces the GrandState::get_state method with two separate
methods, current_state and parent_state. This change improves
separation of concerns and makes the code easier to understand.

The yielded values of the trap::Iter iterator are also changed to
(&Condition, &TrapState, Option<&TrapState>) to reflect the fact that
the current state is always available, even if it was not set by the
user.