wait: Don't treat suspended jobs as terminated jobs
When waiting for a suspended job, the built-in now waits indefinitely
until the job is resumed and finished. Currently, there is no way to
cancel the wait. This new behavior is in line with POSIX.1-2024.

The previous behavior may be re-implemented as an option in the future.

Fixes https://github.com/magicant/yash-rs/issues/459