Sort signals in output of `trap -p` and `kill -l`
This commit changes the order of signals when the trap and kill built-
ins list signals. The signals are now sorted in the ascending order of
their numbers. This is aligned with the behavior of other shells,
although the POSIX standard does not specify the order of signals.

Additionally, the trap built-in now deduplicates signals when listing
them. This makes sure that the same signal is not listed multiple times
when it has multiple names. For example, SIGIOT is a synonym for SIGABRT
on Linux, and the trap built-in now lists only SIGABRT, only once.