Update comments on option portability
We will be providing an error message like "the -x option is not
supported in the POSIX mode" when:

- The user specifies an option that is not supported in the POSIX mode.
- The user uses a long option in the POSIX mode.

For this message to be valid, the flags in the `Mode` struct should be
renamed in terms of POSIX compatibility. This commit updates the
comments in the `Mode` struct to reflect this.