Update trick to work around macOS kill(2) issue
This commit updates the trick used to work around the issue with kill(2)
on macOS. The previous trick simply used a subshell to call the kill
built-in. The new trick uses dummy sleep commands to generate more
SIGCHLD signals and thereby increase the likelihood that the signal sent
by the kill built-in is delivered to the shell before the subshell exits.