Ensure the shell is in the foreground when job control is enabled
This commit implements part of the initialization of the job-control
shell that is specified in POSIX.1-2024. We do it slightly differently
from the standard due to the reason described in the comment for the
Env::ensure_foreground method.

This commit also ensures that the shell is in the foreground when job
control is enabled after the shell starts.

Fixes https://github.com/magicant/yash-rs/issues/481
Fixes https://github.com/magicant/yash-rs/issues/421

This commit also updates the test harness to use `tcgetsid` instead of
`tcgetpgrp` to check the session ID of the terminal. It is unclear
whether the current process becomes the foreground process group when it
obtains the controlling terminal, so we should not rely on the initial
foreground process group being the same as the shell's process group,
just in case.