Reimport job-y.sh test file
This commit adds another invocation of `fg`, in which the job is
suspended, to the test case "discard remaining commands when a command
suspends". Previously, the test case only covered the behavior when an
initially foreground job is suspended. This commit adds coverage for the
behavior when a resumed job is suspended.

This commit also adds a workaround for macOS, where the test case is
kind of flaky.

The file is synced with the upstream repository:
https://github.com/magicant/yash/blob/5f55049e7c204666878d8c3e34da080b16324044/tests/job-y.tst