Remove unnecessary tcsetpgrp_without_block calls
We make sure that the shell is in the foreground process group when the
shell starts job control. This means that we can safely assume that the
shell is in the foreground process group when executing a pipeline or
bringing a job to the foreground in the `fg` built-in.

Fixes https://github.com/magicant/yash-rs/issues/485