Support expected failures in tests
This commit imports the functionality to mark tests as expected failures
in the scripted test suite. This is done by adding the "-f" option to
the "testcase" function in the "run-test.sh" script. The expected
failures are then reported as "OK[FAILED_AS_EXPECTED]" in the test
results. This allows us to keep track of tests that are known to fail
without cluttering the test results with failures.
https://github.com/magicant/yash/commit/16c61e88cdbd2e8677ba817abaa82de88d172851

Redirection delimiters are rewritten as quoted strings instead of
backslash-escaped strings. This renders better syntax highlighting in
VS Code.