Force ChildProcessTask to terminate itself
POSIX.1-2024 requires that the shell kills itself with a signal if the
exit status of the shell indicates that the last command terminated
with a signal. This means that the shell should no longer naively exit
with the exit status of the last command, but instead should examine the
exit status and kill itself with the signal if necessary. To make sure
that this is done correctly, the `ChildProcessTask` type is updated to
return a `Future` with an `Infallible` output type. This way, the task
is forced to terminate itself with a signal or exit cleanly.