Suppress auto-exec optimization in exit status test
This commit updates a test case in which a subshell is expected to kill
itself with SIGTERM when reaching the end of the script. The test is
succeeding and actually testing the expected behavior of the shell.
However, the test may not be testing the expected behavior in the future
if the shell is optimized to avoid forking a subshell when executing the
last command in a subshell. This commit adds a dummy trap to suppress
the auto-exec optimization in the test case. This change ensures that the
test remains valid regardless of future optimizations.