Don't export variables that contain '=' in their name
According to POSIX.1-2024 ([[1]]):

> If more than one string in an environment of a process has the same
> name, the consequences are undefined.

So we must avoid exporting more than one variable with the same name.
Previously, the `env_c_strings` function naively converted all variables
to C strings, but that could lead to more than one variable being
exported with the same name if the variable name contained a `=`
character. For example, if one defined two variables named `foo` and
`foo=bar`, both the variable would be interpreted to have the name `foo`
because the first `=` is always regarded as the name-value separator.

To prevent this issue, the `env_c_strings` function now ignores
variables with names containing `=`.

[[1]]: https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/V1_chap08.html#tag_08_01