Don't print variables with '=' in their names in typeset built-in
Because of the two previous commits, users are no longer possible to
create variables with `=` in their names. However, it would be more
preferable to handle such variables gracefully when printing them in the
`typeset`, `export`, and `readonly` built-ins, just in case the
built-ins are used with a `VariableSet` that includes such badly named
variables.

The output of these built-ins should be formatted so that, if the output
is re-evaluated as a shell script, the shell should restore the variable
definitions recorded in the built-in's output. Since the `=` character
is used as a separator between the variable name and value, the name
cannot contain the `=` character. This commit changes the
`PrintVariables::execute` implementation so that such variables are
ignored.