Don't print variables that don't have a valid name in set built-in
The output of the `set` built-in should be able to be parsed and
executed as a shell script to restore the variable definitions. For the
script to parse correctly, assignments in the script must have a valid
variable name. This commit modifies the built-in so that it does not
print variables that do not have a valid name in the output.

To allow the built-in to check if a variable has a valid name, this
commit also adds utility functions in the yash-syntax crate.